refactor(cosmos-data-layer): moved types to data dir
